usr/backend/requirements.txt
Django==4.2.7
mongoengine==0.27.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.61.1
google-auth==2.25.1
firebase-admin==6.2.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
drf-yasg==1.21.7
drf-spectacular==0.27.0
pytest==7.4.3
pytest-django==4.7.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
